+++ server/src/main/java/org/allaymc/server/command/defaults/GCCommand.java
@@ -3,6 +3,7 @@ package org.allaymc.server.command.defaults;
 import org.allaymc.api.command.SimpleCommand;
 import org.allaymc.api.command.tree.CommandTree;
 import org.allaymc.api.i18n.TrKeys;
+import org.allaymc.api.server.Server;
 
 import static java.lang.Runtime.getRuntime;
 import static org.allaymc.api.math.MathUtils.round;
@@ -21,6 +22,11 @@ public class GCCommand extends SimpleCommand {
                 .exec(context -> {
                     var memory = getCurrentMemoryUsage();
                     System.gc();
+                    for (var world : Server.getInstance().getWorldPool().getWorlds().values()) {
+                        for (var dimension : world.getDimensions().values()) {
+                            dimension.getChunkService().removeUnusedChunksImmediately();
+                        }
+                    }
                     var freedMemory = memory - getCurrentMemoryUsage();
                     context.getSender().sendTr(TrKeys.A_COMMAND_GC_COMPLETED, freedMemory);
                     return context.success();