+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Unless otherwise specified, any version comparison below is the comparison of se
   and sources can be registered by plugin now.
 - (API) Added `Form#onClose(Consumer<ModalFormCancelReason>)` method that can be used to set a callback which will be called
   with the close reason when the form is closed. The old `Form#onClose` method is still available that just ignores the reason.
+- (API) Added `ChunkService#removeUnusedChunksImmediately` method that can remove unused chunks immediately. Also, the `/gc` command
+  will call this method in all dimensions now.
 
 ### Changed